Improvements to the template display:
1) It now deals better with contigs that do not start as base 1.

2) The depth plot should no longer change Y values so drastically when
zooming in and out. This still exists in part due to aliasing, but
it's far better than it was.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2442 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/gap5/gap_range.h
@@ -30,8 +30,8 @@ typedef struct {
 } gap_filter_t;
 
 typedef struct {
-    int t;
-    int s;
+    double t;
+    double s;
 } gap_depth_t;
